--- output/day_3_1.txt
@@ -1,1 +1,0 @@
-1
